requirements.txt
Flask
pypdf
python-docx
google-generativeai
gunicorn